+++ src/CoreEx/Invokers/InvokeArgs.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/CoreEx
 
+using CoreEx.Abstractions;
 using CoreEx.Configuration;
 using CoreEx.Results;
+using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using System;
@@ -15,19 +17,20 @@ namespace CoreEx.Invokers
     /// </summary>
     public struct InvokeArgs
     {
-        private static readonly ConcurrentDictionary<Type, (ActivitySource ActivitySource, bool IsTracingEnabled, bool IsLoggingEnabled)> _invokerOptions = new();
+        private static readonly ConcurrentDictionary<Type, ActivitySource> _activitySources = new();
 
         private const string NullName = "null";
-        private const string InvokerType = "invoker.type";
-        private const string InvokerOwner = "invoker.owner";
-        private const string InvokerMember = "invoker.member";
-        private const string InvokerResult = "invoker.result";
-        private const string InvokerFailure = "invoker.failure";
-        private const string CompleteState = "Complete";
-        private const string SuccessState = "Success";
-        private const string FailureState = "Failure";
-        private const string ExceptionState = "Exception";
+        private const string InvokerTypeName = "invoker.type";
+        private const string InvokerOwnerName = "invoker.owner";
+        private const string InvokerMemberName = "invoker.member";
+        private const string InvokerResultName = "invoker.result";
+        private const string InvokerFailureName = "invoker.failure";
+        private const string CompleteStateText = "Complete";
+        private const string SuccessStateText = "Success";
+        private const string FailureStateText = "Failure";
+        private const string ExceptionStateText = "Exception";
 
+        private Type? _ownerType;
         private bool _isComplete;
 
         /// <summary>
@@ -55,51 +58,75 @@ namespace CoreEx.Invokers
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="InvokeArgs"/> struct.
+        /// Provides the default <see cref="IInvoker.CallerLoggerFormatter"/> implementation.
         /// </summary>
-        public InvokeArgs() => Type = typeof(Invoker);
+        /// <param name="args">The <see cref="InvokeArgs"/>.</param>
+        /// <returns>The caller information to be included in the log output.</returns>
+        public static string DefaultCallerLogFormatter(InvokeArgs args) => args.OwnerType is null ? args.MemberName ?? NullName : $"{args.OwnerType}->{args.MemberName ?? NullName}";
+
+        /// <summary>
+        /// Gets or sets the <see cref="ICacheEntry.AbsoluteExpirationRelativeToNow"/> <see cref="TimeSpan"/> for <i>tracing</i> and <i>logging</i> enablement determination.
+        /// </summary>
+        /// <remarks>These are cached to avoid the overhead of repeated configuration lookups and allow for dynamic configuration changes.</remarks>
+        public static TimeSpan AbsoluteExpirationTimeSpan { get; set; } = TimeSpan.FromMinutes(5);
 
         /// <summary>
         /// Initializes a new instance of the <see cref="InvokeArgs"/> struct.
         /// </summary>
-        /// <param name="invokerType">The invoker <see cref="Type"/> used to manage the activity sources.</param>
-        /// <param name="ownerType">The invoking (owner) <see cref="Type"/> to include as part of the <see cref="Activity.OperationName"/>.</param>
+        /// <remarks>This will throw a <see cref="NotSupportedException"/>.</remarks>
+        public InvokeArgs() => throw new NotSupportedException($"The {nameof(InvokeArgs)} default constructor is not supported; please use other(s).");
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="InvokeArgs"/> struct.
+        /// </summary>
+        /// <param name="invoker">The initiating <see cref="IInvoker"/>.</param>
+        /// <param name="owner">The invoking (owner) value.</param>
         /// <param name="memberName">The calling member name.</param>
         /// <param name="invokeArgs">The optional parent <see cref="InvokeArgs"/>.</param>
-        /// <remarks>Creates the tracing <see cref="Activity.OperationName"/> by concatenating the invoking <paramref name="ownerType"/> (<see cref="Type.FullName"/>) and <paramref name="memberName"/> separated by '<c> -> </c>'. This is <i>not</i>
+        /// <remarks>Creates the tracing <see cref="Activity.OperationName"/> by concatenating the invoking <see name="OwnerType"/> (<see cref="Type.FullName"/>) and <paramref name="memberName"/> separated by '<c> -> </c>'. This is <i>not</i>
         /// meant to represent the fully-qualified member/method name.</remarks>
-        public InvokeArgs(Type invokerType, Type? ownerType, string? memberName, InvokeArgs? invokeArgs)
+        public InvokeArgs(IInvoker invoker, object? owner, string? memberName, InvokeArgs? invokeArgs)
         {
-            Type = invokerType;
-            OwnerType = ownerType;
+            Invoker = invoker ?? throw new ArgumentNullException(nameof(invoker));
+            InvokerType = invoker.GetType();
+            Owner = owner;
             MemberName = memberName;
 
             try
             {
-                var options = _invokerOptions.GetOrAdd(invokerType, type => (new ActivitySource(type.FullName ?? NullName), IsTracingEnabled(invokerType), IsLoggingEnabled(invokerType)));
-                if (options.IsTracingEnabled)
+                var enabled = Internal.MemoryCache.GetOrCreate<(bool IsTracingEnabled, bool IsLoggingEnabled)>(InvokerType, e =>
                 {
-                    Activity = options.ActivitySource.CreateActivity(ownerType is null ? memberName ?? NullName : $"{ownerType}->{memberName ?? NullName}", ActivityKind.Internal);
+                    // These are cached to avoid the overhead of repeated configuration lookups and allow for dynamic configuration changes.
+                    var type = (Type)e.Key;
+                    e.SetAbsoluteExpiration(TimeSpan.FromMinutes(5));
+                    return (IsTracingEnabled(type), IsLoggingEnabled(type));
+                });
+                
+                if (enabled.IsTracingEnabled)
+                {
+                    var activitySource = _activitySources.GetOrAdd(InvokerType, type => new ActivitySource(type.FullName ?? NullName));
+                    Activity = activitySource.CreateActivity(OwnerType is null ? memberName ?? NullName : $"{OwnerType}->{memberName ?? NullName}", ActivityKind.Internal);
                     if (Activity is not null)
                     {
                         if (invokeArgs.HasValue && invokeArgs.Value.Activity is not null)
                             Activity.SetParentId(invokeArgs.Value.Activity!.TraceId, invokeArgs.Value.Activity.SpanId, invokeArgs.Value.Activity.ActivityTraceFlags);
 
-                        Activity.SetTag(InvokerType, options.ActivitySource!.Name);
-                        Activity.SetTag(InvokerOwner, ownerType?.FullName);
-                        Activity.SetTag(InvokerMember, memberName);
+                        Activity.SetTag(InvokerTypeName, activitySource.Name);
+                        Activity.SetTag(InvokerOwnerName, OwnerType?.FullName);
+                        Activity.SetTag(InvokerMemberName, memberName);
+                        Invoker.OnActivityStart?.Invoke(this);
                         Activity.Start();
                     }
                 }
 
-                if (options.IsLoggingEnabled)
+                if (enabled.IsLoggingEnabled)
                 {
                     Logger = ExecutionContext.GetService<ILogger<Invoker>>();
                     if (Logger is null || !Logger.IsEnabled(LogLevel.Debug))
                         Logger = null;
                     else
                     {
-                        Logger.LogDebug("{InvokerType}: Start {InvokerCaller}.", invokerType.ToString(), FormatCaller());
+                        Logger.LogDebug("{InvokerType}: Start {InvokerCaller}.", InvokerType.ToString(), Invoker.CallerLoggerFormatter(this));
                         Stopwatch = Stopwatch.StartNew();
                     }
                 }
@@ -113,14 +140,24 @@ namespace CoreEx.Invokers
         }
 
         /// <summary>
-        /// Gets the invoker <see cref="Type"/>.
+        /// Gets the initiating <see cref="IInvoker"/>.
         /// </summary>
-        public Type Type { get; }
+        public IInvoker Invoker { get; }
 
         /// <summary>
-        /// Gets the invoking (owner) <see cref="Type"/>
+        /// Gets the <see cref="IInvoker"/> <see cref="Type"/>.
         /// </summary>
-        public Type? OwnerType { get; }
+        public Type InvokerType { get; }
+
+        /// <summary>
+        /// Gets the owning invocation value.
+        /// </summary>
+        public object? Owner { get; }
+
+        /// <summary>
+        /// Gets the owning invocation <see cref="Type"/>
+        /// </summary>
+        public Type? OwnerType => _ownerType ??= Owner?.GetType();
 
         /// <summary>
         /// Gets the calling member name.
@@ -144,11 +181,6 @@ namespace CoreEx.Invokers
         public Stopwatch? Stopwatch { get; }
 
         /// <summary>
-        /// Formats the caller.
-        /// </summary>
-        private readonly string FormatCaller() => OwnerType is null ? MemberName ?? NullName : $"{OwnerType}->{MemberName ?? NullName}";
-
-        /// <summary>
         /// Adds the result outcome to the <see cref="Activity"/> (where started).
         /// </summary>
         /// <typeparam name="TResult">The result <see cref="Type"/>.</typeparam>
@@ -160,10 +192,11 @@ namespace CoreEx.Invokers
             if (Activity is not null)
             {
                 var ir = result as IResult;
-                Activity.SetTag(InvokerResult, ir is null ? CompleteState : (ir.IsSuccess ? SuccessState : FailureState));
+                Activity.SetTag(InvokerResultName, ir is null ? CompleteStateText : (ir.IsSuccess ? SuccessStateText : FailureStateText));
                 if (ir is not null && ir.IsFailure)
-                    Activity.SetTag(InvokerFailure, $"{ir.Error.Message} ({ir.Error.GetType()})");
+                    Activity.SetTag(InvokerFailureName, $"{ir.Error.Message} ({ir.Error.GetType()})");
 
+                Invoker.OnActivityComplete?.Invoke(this);
                 _isComplete = true;
             }
 
@@ -172,7 +205,7 @@ namespace CoreEx.Invokers
                 Stopwatch!.Stop();
                 var ir = result as IResult;
                 Logger.LogDebug("{InvokerType}: {InvokerResult} {InvokerCaller}{InvokerFailure} [{Elapsed}ms].", 
-                    Type.ToString(), ir is null ? CompleteState : (ir.IsSuccess ? SuccessState : FailureState), FormatCaller(), (ir is not null && ir.IsFailure) ? $" {ir.Error.Message} ({ir.Error.GetType()})" : string.Empty, Stopwatch.Elapsed.TotalMilliseconds);
+                    InvokerType.ToString(), ir is null ? CompleteStateText : (ir.IsSuccess ? SuccessStateText : FailureStateText), Invoker.CallerLoggerFormatter(this), (ir is not null && ir.IsFailure) ? $" {ir.Error.Message} ({ir.Error.GetType()})" : string.Empty, Stopwatch.Elapsed.TotalMilliseconds);
                
                 _isComplete = true;
             }
@@ -181,22 +214,23 @@ namespace CoreEx.Invokers
         }
 
         /// <summary>
-        /// Completes the <see cref="Activity"/> tracing (where started) recording the <see cref="InvokerResult"/> with the <see cref="ExceptionState"/> and capturing the corresponding <see cref="Exception.Message"/>.
+        /// Completes the <see cref="Activity"/> tracing (where started) recording the <see cref="InvokerResultName"/> with the <see cref="ExceptionStateText"/> and capturing the corresponding <see cref="Exception.Message"/>.
         /// </summary>
         /// <param name="ex">The <see cref="System.Exception"/>.</param>
         public void TraceException(Exception ex)
         {
             if (Activity is not null && ex is not null)
             {
-                Activity.SetTag(InvokerResult, ExceptionState);
-                Activity.SetTag(InvokerFailure, $"{ex.Message} ({ex.GetType()})");
+                Activity.SetTag(InvokerResultName, ExceptionStateText);
+                Activity.SetTag(InvokerFailureName, $"{ex.Message} ({ex.GetType()})");
+                Invoker.OnActivityException?.Invoke(this, ex);
                 _isComplete = true;
             }
 
             if (Logger is not null && ex is not null)
             {
                 Stopwatch!.Stop();
-                Logger.LogDebug("{InvokerType}: {InvokerResult} {InvokerCaller}{InvokerFailure} [{Elapsed}ms].", Type.ToString(), ExceptionState, FormatCaller(), $" {ex.Message} ({ex.GetType()})", Stopwatch.Elapsed.TotalMilliseconds);
+                Logger.LogDebug("{InvokerType}: {InvokerResult} {InvokerCaller}{InvokerFailure} [{Elapsed}ms].", InvokerType.ToString(), ExceptionStateText, Invoker.CallerLoggerFormatter(this), $" {ex.Message} ({ex.GetType()})", Stopwatch.Elapsed.TotalMilliseconds);
                 _isComplete = true;
             }
         }
@@ -204,14 +238,17 @@ namespace CoreEx.Invokers
         /// <summary>
         /// Completes (stops) the <see cref="Activity"/> tracing (where started).
         /// </summary>
-        /// <remarks>Where not previously recorded as complete will set the <see cref="InvokerResult"/> to <see cref="ExceptionState"/>.</remarks>
+        /// <remarks>Where not previously recorded as complete will set the <see cref="InvokerResultName"/> to <see cref="ExceptionStateText"/>.</remarks>
         public readonly void TraceComplete()
         {
             if (Activity is not null)
             {
                 // Where no result then it can only be as a result of an exception.
                 if (!_isComplete)
-                    Activity.SetTag(InvokerResult, ExceptionState);
+                {
+                    Activity.SetTag(InvokerResultName, ExceptionStateText);
+                    Invoker.OnActivityException?.Invoke(this, new InvalidOperationException("The invocation was not completed successfully."));
+                }
 
                 Activity.Stop();
             }
@@ -219,28 +256,28 @@ namespace CoreEx.Invokers
             if (Logger is not null && !_isComplete)
             {
                 Stopwatch!.Stop();
-                Logger.LogDebug("{InvokerType}: {InvokerResult} {InvokerCaller} [{Elapsed}ms].", Type.ToString(), ExceptionState, FormatCaller(), Stopwatch.Elapsed.TotalMilliseconds);
+                Logger.LogDebug("{InvokerType}: {InvokerResult} {InvokerCaller} [{Elapsed}ms].", InvokerType.ToString(), ExceptionStateText, Invoker.CallerLoggerFormatter(this), Stopwatch.Elapsed.TotalMilliseconds);
             }
         }
 
         /// <summary>
         /// Creates (and started) a new <see cref="InvokeArgs"/> instance for a related invocation.
         /// </summary>
-        /// <param name="invokerType">The invoker <see cref="Type"/> used to manage the activity sources.</param>
-        /// <param name="ownerType">The invoking (owner) <see cref="Type"/> to include as part of the <see cref="Activity.OperationName"/>.</param>
+        /// <param name="invoker">The invoker used to manage the activity sources.</param>
+        /// <param name="owner">The invoking (owner) value.</param>
         /// <param name="memberName">The calling member name.</param>
         /// <returns>The <see cref="InvokeArgs"/>.</returns>
-        public readonly InvokeArgs StartNewRelated(Type invokerType, Type? ownerType, string? memberName) => new(invokerType, ownerType, memberName, this);
+        public readonly InvokeArgs StartNewRelated(IInvoker invoker, object? owner, string? memberName) => new(invoker, owner, memberName, this);
 
         /// <summary>
         /// Releases (disposes) all <see cref="ActivitySource"/> instances.
         /// </summary>
         public static void ReleaseAll()
         {
-            foreach (var item in _invokerOptions.ToArray())
+            foreach (var item in _activitySources.ToArray())
             {
-                if (_invokerOptions.TryRemove(item.Key, out var options))
-                    options.ActivitySource?.Dispose();
+                if (_activitySources.TryRemove(item.Key, out var activitySource))
+                    activitySource?.Dispose();
             }
         }
     }